+++ tests/2D_checkNeumannWactivation/test.py
@@ -17,7 +17,7 @@ def mesh(box, meshDen=1):
         cell_type=cell_type
         #variant="zigzag",  # or "up", "down", "center"
     )
-    cells = cells.astype( int )
+    cells = cells.astype( np.uint32 )
     return points, cells, cell_type
 
 def specActivate( p ):
